Migrate fitting helper functions into new fitting_helpers.utils.py module
